creating cart

+++ src/main/java/com/api/zorvanz/domain/order_details/OrderDetailData.java
@@ -1,11 +1,13 @@
 package com.api.zorvanz.domain.order_details;
 
+import java.math.BigDecimal;
+
 public record OrderDetailData(
         Long orderDetailId,
         Long orderId,
         Long productId,
         Integer quantity,
-        Double unitPrice,
-        Double totalAmount
+        BigDecimal unitPrice,
+        BigDecimal totalAmount
 ) {
 }
